feat(devices): add new fn to DriverRegion struct impl and refactor ns16550 and clint0 init fn to avoid code duplication
